client/requirements.txt
numpy==1.26.4
Pillow==10.2.0
plotly==5.20.0
Requests==2.31.0
scipy==1.13.0
streamlit==1.33.0
torch==2.2.0
transformers==4.39.3
